requirements.txt
pytest
selenium
webdriver-manager
pytest-rerunfailures
pytest-pyreport
allure-pytest